requirements.txt
django>=5.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
openai>=1.0.0
django-cors-headers>=4.3.0
django-storages>=1.14.2
boto3>=1.28.0
qrcode>=7.4.2
drf-yasg>=1.21.7
